# okrTrackingSystem/api-docs.yaml
openapi: 3.1.0
info:
  title: OKR Tracking System API
  description: |
    ## OKR Tracking System — Garant Bank

    REST API for managing **Objectives & Key Results (OKR)**, organizational structure, user management, and multi-source performance evaluations.

    ### Key Features
    - **OKR Management** — Create and track Objectives with weighted Key Results
    - **Multi-Source Evaluations** — Director (star rating), HR (letter grade), and Business Block assessments
    - **Score Calculation** — Automatic weighted scoring: 60% OKR + 20% Director + 20% HR
    - **Organizational Hierarchy** — Divisions → Departments → Objectives → Key Results
    - **Role-Based Access Control** — 6 roles with fine-grained permissions
    - **Excel Export** — Export all OKR data to Excel spreadsheets

    ### Authentication
    All endpoints (except login and export) require a **JWT Bearer token**.
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click the **Authorize** button above and enter the token

    ### Roles
    | Role | Description |
    |------|-------------|
    | `ADMIN` | Full system access, user management |
    | `DIRECTOR` | Division oversight, star-based evaluations |
    | `DEPARTMENT_LEADER` | Department OKR management |
    | `HR` | Letter-grade evaluations (A/B/C/D) |
    | `BUSINESS_BLOCK` | Business block star-based evaluations |
    | `EMPLOYEE` | View own profile and assigned departments |
  version: 1.0.0
...
servers:
...
security:
- Bearer Authentication: []
...
paths:
...
  /api/divisions/{id}:
    get:
      tags:
      - Divisions
      summary: Get division by ID
      description: Returns a single division with its leader info and department summaries.
      operationId: getDivisionById
      parameters:
      - name: id
        in: path
        description: Division ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Division found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DivisionDTO"
        "404":
          description: Division not found
    put:
      tags:
      - Divisions
      summary: Update division
      description: "Update a division's name and/or leader. Requires edit permission\
        \ (ADMIN, DIRECTOR, or division leader)."
      operationId: updateDivision
      parameters:
      - name: id
        in: path
        description: Division ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDivisionRequest"
        required: true
      responses:
        "200":
          description: Division updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DivisionDTO"
        "400":
          description: Invalid request
        "403":
          description: No edit permission for this division
        "404":
          description: Division not found
    delete:
      tags:
      - Divisions
      summary: Delete division
      description: Delete a division. The division must have no departments assigned.
        **Requires ADMIN role.**
      operationId: deleteDivision
      parameters:
      - name: id
        in: path
        description: Division ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Division deleted
        "400":
          description: Division still has departments
        "403":
          description: Only ADMIN can delete divisions
        "404":
          description: Division not found
...
  /api/divisions:
    get:
      tags:
      - Divisions
      summary: Get all divisions
      description: Returns all divisions with their leader info and department summaries.
        Accessible to all authenticated users.
      operationId: getAllDivisions
      responses:
        "200":
          description: List of all divisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DivisionDTO"
    post:
      tags:
      - Divisions
      summary: Create division
      description: Create a new division. Optionally assign a division leader. **Requires
        ADMIN or DIRECTOR role.**
      operationId: createDivision
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDivisionRequest"
        required: true
      responses:
        "201":
          description: Division created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DivisionDTO"
        "400":
          description: "Invalid request (e.g., name too short)"
        "403":
          description: Insufficient permissions
...
  /api/divisions/{id}/score:
    get:
      tags:
      - Divisions
      summary: Get division score
      description: "Returns the division's aggregated score calculated from its departments'\
        \ scores, including score level and color classification."
      operationId: getDivisionScore
      parameters:
      - name: id
        in: path
        description: Division ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Division score details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DivisionWithScoreDTO"
        "404":
          description: Division not found
  /api/divisions/{id}/departments:
    get:
      tags:
      - Divisions
      summary: Get departments in division
      description: Returns a summary list of all departments belonging to the specified
        division.
      operationId: getDepartmentsByDivision
      parameters:
      - name: id
        in: path
        description: Division ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: List of departments in the division
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepartmentSummaryDTO"
        "404":
          description: Division not found
...
  /api/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current user
      description: Returns the profile information of the currently authenticated
        user based on the JWT token.
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    DepartmentSummaryDTO:
      type: object
      description: Lightweight department reference
      properties:
        id:
          type: string
          description: Department ID
          example: dept-001
        name:
          type: string
          description: Department name
          example: Software Development
    UserDTO:
      type: object
      description: User information (password excluded)
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Unique username
          example: john.doe
        email:
          type: string
          description: Email address
          example: john.doe@garantbank.uz
        fullName:
          type: string
          description: Full display name
          example: John Doe
        role:
          type: string
          description: User role
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: EMPLOYEE
        profilePhotoUrl:
          type: string
          description: URL to user's profile photo
          example: /uploads/photo-abc123.jpg
        jobTitle:
          type: string
          description: Job title
          example: Senior Developer
        phoneNumber:
          type: string
          description: Phone number
          example: +998 90 123 45 67
        bio:
          type: string
          description: User bio/description
          example: Experienced developer focused on backend systems
        lastLogin:
          type: string
          format: date-time
          description: Timestamp of last login
          example: 2025-01-15T10:30:00
        assignedDepartments:
          type: array
          description: Departments the user is assigned to
          items:
            $ref: "#/components/schemas/DepartmentSummaryDTO"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: 2025-01-01T09:00:00
        active:
          type: boolean
        isActive:
          type: boolean
          description: Whether the user account is active
          example: true
        canEditAssignedDepartments:
          type: boolean
          description: Whether the employee can edit their assigned departments
          example: false
        readOnly:
          type: boolean
          description: Whether the user is in read-only mode (cannot make any edits)
          example: false
...
    UpdateDivisionRequest:
      type: object
      description: Request to update a division (all fields optional)
      properties:
        name:
          type: string
          description: New division name (2-100 characters)
          example: Information Technology & Security
          maxLength: 100
          minLength: 2
        leaderId:
          type: string
          format: uuid
          description: New division leader user ID
          example: 550e8400-e29b-41d4-a716-446655440000
    DivisionDTO:
      type: object
      description: Division with leader info and department summaries
      properties:
        id:
          type: string
          description: Division ID
          example: div-001
        name:
          type: string
          description: Division name
          example: Information Technology
        divisionLeader:
          $ref: "#/components/schemas/UserSummaryDTO"
          description: Division leader (basic info)
        departments:
          type: array
          description: Departments in this division
          items:
            $ref: "#/components/schemas/DepartmentSummaryDTO"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: 2025-01-01T09:00:00
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: 2025-01-15T14:30:00
    UserSummaryDTO:
      type: object
      description: Lightweight user summary for display purposes
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Username
          example: john.doe
        fullName:
          type: string
          description: Full name
          example: John Doe
        profilePhotoUrl:
          type: string
          description: Profile photo URL
          example: /uploads/photo-abc.jpg
...
    CreateDivisionRequest:
      type: object
      description: Request to create a new division
      properties:
        name:
          type: string
          description: Division name (2-100 characters)
          example: Information Technology
          maxLength: 100
          minLength: 2
        leaderId:
          type: string
          format: uuid
          description: Optional ID of the user to assign as division leader
          example: 550e8400-e29b-41d4-a716-446655440000
      required:
      - name
...
    DivisionWithScoreDTO:
      type: object
      description: Division with aggregated score from its departments
      properties:
        id:
          type: string
          description: Division ID
          example: div-001
        name:
          type: string
          description: Division name
          example: Information Technology
        score:
          type: number
          format: double
          description: Aggregated score (0.0-1.0)
          example: 0.65
        scoreLevel:
          type: string
          description: Score level classification
          enum:
          - не_соответствует
          - ниже_ожиданий
          - на_уровне_ожиданий
          - превышает_ожидания
          - исключительно
          example: на_уровне_ожиданий
        color:
          type: string
          description: Hex color for score level
          example: "#4CAF50"
        percentage:
          type: number
          format: double
          description: Score as percentage (0-100)
          example: 67.5
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter the JWT token obtained from `POST /api/auth/login`
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT